refactor(model): simplify Card model using data class and static all cards, remove Card subclasses
