+++ src/main/kotlin/blackjack/model/Suit.kt
@@ -1,0 +1,8 @@
+package blackjack.model
+
+enum class Suit(val symbol: String) {
+    DIAMONDS("♦"),
+    HEARTS("♥"),
+    CLUBS("♣"),
+    SPADES("♠"),
+}
